separate cointoss modules
